+++ Rainbow_NLP/Extention/UIColor + ext.swift
@@ -1,0 +1,26 @@
+//
+//  UIColor + ext.swift
+//  Rainbow_NLP
+//
+//  Created by Павел Грицков on 16.11.23.
+//
+
+import UIKit
+
+extension UIColor {
+    static var grayRainbowButton: UIColor {
+        UIColor(red: 0.135, green: 0.169, blue: 0.153, alpha: 1)
+    }
+    static var downButtons: UIColor {
+        UIColor(red: 0.43, green: 0.04, blue: 0.44, alpha: 1)
+    }
+    static var newGameButtonColor: UIColor {
+        UIColor(red:  0.871, green: 0.132, blue: 0.132, alpha: 1)
+    }
+    static var newStatsButtonColor: UIColor {
+        UIColor(red: 0.19, green: 0.65, blue: 0.29, alpha: 1)
+    }
+    static var backgroundColor: UIColor {
+        UIColor(red: 0, green: 0, blue: 0, alpha: 0.2)
+    }
+}
