+++ Rainbow_NLP/MainView.swift
@@ -8,14 +8,19 @@
 import UIKit
 
 class MainView: UIView {
+    private struct Drawing {
+        static let height: CGFloat = 50.0
+    }
     
-    // MARK: - Elements
-    let nlpLabel = MainView.makeNlpLabel()
-    let rainbowImage = MainView.makeRainbowImage()
-    let rainbowLabel = MainView.makeRainbowLabel()
+    // MARK: - Private UI
+    private let nlpLabel = MainView.makeNlpLabel()
+    private let rainbowImage = MainView.makeRainbowImage()
+    private let rainbowLabel = MainView.makeRainbowLabel()
+    
+    // MARK: - Public UI
     let newGameButton = MainView.makeNewGameButton()
-    let statsButton = MainView.makeNewStatsButton()
     let settingButton = MainView.makeNewSettingsButton()
+    let statsButton = MainView.makeNewStatsButton()
     let rulesButton = MainView.makeNewRulesButton()
     
     // MARK: - init(_:)
@@ -33,50 +38,53 @@ class MainView: UIView {
         )
     }
     
-    // MARK: - Constraints
+    // MARK: - Life cycle
     override func layoutSubviews() {
         super.layoutSubviews()
         setLayout()
     }
+    
+    // MARK: - Private Methods
     private func setLayout() {
-        NSLayoutConstraint.activate([rainbowImage.topAnchor.constraint(equalTo: topAnchor, constant: K.heigh() / 2),
-                                     rainbowImage.centerXAnchor.constraint(equalTo: centerXAnchor),
-                                     rainbowImage.widthAnchor.constraint(equalToConstant: 302),
-                                     rainbowImage.heightAnchor.constraint(equalToConstant: 150),
-                                     
-                                     newGameButton.centerXAnchor.constraint(equalTo: centerXAnchor),
-                                     newGameButton.widthAnchor.constraint(equalToConstant: 274),
-                                     newGameButton.heightAnchor.constraint(equalToConstant: 83),
-                                     newGameButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 53),
-                                     newGameButton.centerYAnchor.constraint(equalTo: centerYAnchor, constant: 134.5),
-                                     
-                                     statsButton.centerXAnchor.constraint(equalTo: centerXAnchor),
-                                     statsButton.widthAnchor.constraint(equalToConstant: 274),
-                                     statsButton.heightAnchor.constraint(equalToConstant: 83),
-                                     statsButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 53),
-                                     statsButton.centerYAnchor.constraint(equalTo: centerYAnchor, constant: 236.5),
-                                     
-                                     settingButton.widthAnchor.constraint(equalToConstant: Drawing.height),
-                                     settingButton.heightAnchor.constraint(equalToConstant: Drawing.height),
-                                     settingButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 28),
-                                     settingButton.bottomAnchor.constraint(equalTo: self.safeAreaLayoutGuide.bottomAnchor, constant: 0),
-                                     
-                                     rulesButton.widthAnchor.constraint(equalToConstant: 45),
-                                     rulesButton.heightAnchor.constraint(equalToConstant: Drawing.height),
-                                     rulesButton.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -26),
-                                     rulesButton.bottomAnchor.constraint(equalTo: self.safeAreaLayoutGuide.bottomAnchor, constant: 0),
-                                     
-                                     rainbowLabel.heightAnchor.constraint(equalToConstant: 131),
-                                     rainbowLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-                                     rainbowLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-                                     rainbowLabel.topAnchor.constraint(equalTo: topAnchor, constant: 238),
-                                     rainbowLabel.bottomAnchor.constraint(equalTo: topAnchor, constant: 409),
-                                     
-                                     nlpLabel.heightAnchor.constraint(equalToConstant: 71),
-                                     nlpLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-                                     nlpLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-                                     nlpLabel.topAnchor.constraint(equalTo: topAnchor, constant: 227)
-                                    ])
+        NSLayoutConstraint.activate(
+            [rainbowImage.topAnchor.constraint(equalTo: topAnchor, constant: K.heigh() / 2),
+             rainbowImage.centerXAnchor.constraint(equalTo: centerXAnchor),
+             rainbowImage.widthAnchor.constraint(equalToConstant: 302),
+             rainbowImage.heightAnchor.constraint(equalToConstant: 150),
+             
+             newGameButton.centerXAnchor.constraint(equalTo: centerXAnchor),
+             newGameButton.widthAnchor.constraint(equalToConstant: 274),
+             newGameButton.heightAnchor.constraint(equalToConstant: 83),
+             newGameButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 53),
+             newGameButton.centerYAnchor.constraint(equalTo: centerYAnchor, constant: 134.5),
+             
+             statsButton.centerXAnchor.constraint(equalTo: centerXAnchor),
+             statsButton.widthAnchor.constraint(equalToConstant: 274),
+             statsButton.heightAnchor.constraint(equalToConstant: 83),
+             statsButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 53),
+             statsButton.centerYAnchor.constraint(equalTo: centerYAnchor, constant: 236.5),
+             
+             settingButton.widthAnchor.constraint(equalToConstant: Drawing.height),
+             settingButton.heightAnchor.constraint(equalToConstant: Drawing.height),
+             settingButton.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 28),
+             settingButton.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor, constant: 0),
+             
+             rulesButton.widthAnchor.constraint(equalToConstant: 45),
+             rulesButton.heightAnchor.constraint(equalToConstant: Drawing.height),
+             rulesButton.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -26),
+             rulesButton.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor, constant: 0),
+             
+             rainbowLabel.heightAnchor.constraint(equalToConstant: 131),
+             rainbowLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
+             rainbowLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
+             rainbowLabel.topAnchor.constraint(equalTo: topAnchor, constant: 238),
+             rainbowLabel.bottomAnchor.constraint(equalTo: topAnchor, constant: 409),
+             
+             nlpLabel.heightAnchor.constraint(equalToConstant: 71),
+             nlpLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
+             nlpLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
+             nlpLabel.topAnchor.constraint(equalTo: topAnchor, constant: 227)
+            ])
     }
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
@@ -85,11 +93,12 @@ class MainView: UIView {
 extension MainView {
     static func makeRainbowImage() -> UIImageView {
         let rainbowView = UIImageView()
-        rainbowView.image = UIImage(named: "Rainbowimage")
+        rainbowView.image = UIImage(named: "RainbowImage")
         rainbowView.contentMode = .scaleAspectFit
         rainbowView.translatesAutoresizingMaskIntoConstraints = false
         return rainbowView
     }
+    
     static func makeNlpLabel() -> UILabel {
         let nlp = UILabel()
         nlp.font = UIFont(name: "SFProDisplay-Regular", size: 36)
@@ -99,6 +108,7 @@ extension MainView {
         nlp.translatesAutoresizingMaskIntoConstraints = false
         return nlp
     }
+    
     static func makeRainbowLabel() -> UILabel {
         let rainbow = UILabel()
         rainbow.font = UIFont(name: "CormorantInfant-Regular", size: 64)
@@ -111,7 +121,7 @@ extension MainView {
     }
     static func makeNewGameButton() -> UIButton {
         let button = UIButton(type: .system)
-        button.backgroundColor = UIColor.newGameButtonCollor
+        button.backgroundColor = UIColor.newGameButtonColor
         button.tintColor = UIColor.white
         button.layer.cornerRadius = 10
         button.titleLabel?.font = UIFont(name: "SFPro-Regular", size: 20)
@@ -119,6 +129,7 @@ extension MainView {
         button.translatesAutoresizingMaskIntoConstraints = false
         return button
     }
+    
     static func makeNewStatsButton() -> UIButton {
         let button = UIButton(type: .system)
         button.backgroundColor = UIColor.newStatsButtonColor
@@ -129,6 +140,7 @@ extension MainView {
         button.translatesAutoresizingMaskIntoConstraints = false
         return button
     }
+    
     static func makeNewSettingsButton() -> UIButton {
         let button = UIButton(type: .system)
         button.setImage(UIImage(named: "Settings button"), for: .normal)
@@ -136,6 +148,7 @@ extension MainView {
         button.translatesAutoresizingMaskIntoConstraints = false
         return button
     }
+    
     static func makeNewRulesButton() -> UIButton {
         let button = UIButton(type: .system)
         button.setImage(UIImage(named: "Rules button"), for: .normal)
@@ -143,25 +156,4 @@ extension MainView {
         button.translatesAutoresizingMaskIntoConstraints = false
         return button
     }
-}
-extension UIColor {
-    static var grayRainbowButton: UIColor {
-        UIColor(red: 0.135, green: 0.169, blue: 0.153, alpha: 1)
-    }
-    static var downButtons: UIColor {
-        UIColor(red: 0.43, green: 0.04, blue: 0.44, alpha: 1)
-    }
-    static var newGameButtonCollor: UIColor {
-        UIColor(red:  0.871, green: 0.132, blue: 0.132, alpha: 1)
-    }
-    static var newStatsButtonColor: UIColor {
-        UIColor(red: 0.19, green: 0.65, blue: 0.29, alpha: 1)
-    }
-    static var backgroundColor: UIColor {
-        UIColor(red: 0, green: 0, blue: 0, alpha: 0.2)
-    }
-    
-}
-private struct Drawing {
-    static let height: CGFloat = 50.0
 }
